requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.1
cssselect2==0.3.0
defusedxml==0.6.0
Pillow==7.2.0
pycparser==2.20
svgwrite==1.4
tinycss2==1.0.2
webencodings==0.5.1
